+++ package.json
@@ -3,8 +3,8 @@
   "version": "1.0.0",
   "description": "JavaScript test project",
   "scripts": {
-    "test": "ts-node tests/validate-store-update.ts",
-    "write-checksum": "ts-node write-checksum.ts"
+    "test": "ts-node src/validate-store-update.ts",
+    "write-checksum": "ts-node src/write-checksum.ts"
   },
   "devDependencies": {
     "@types/adm-zip": "^0.5.0",
