+++ src/validate-store-update.ts
